embed compiler inside idyll-document, fix documentation and tests
