+++ packages/idyll-docs/pages/docs/publishing/embedding.js
@@ -3,72 +3,6 @@ import markdown from 'markdown-in-js'
 import Layout from '../../../components/layout'
 
 
-const Content = () => markdown`
-
-# Embedding Idyll in an existing web page
-
-The Idyll runtime is available as a React
-component, allowing you to embed interactive
-Idyll content anywhere on the web.
-
-To do this, you must first install the dependencies:
-
-\`\`\`sh
-$ npm i --save idyll-document idyll-compiler idyll-components
-\`\`\`
-
-then, add it to your page. If you are already using React, you
-can include this as a standard component:
-
-\`\`\`javascript
-
-import IdyllDocument from 'idyll-document';
-import compile from 'idyll-compiler';
-import * as components from 'idyll-components';
-
-// An example functional component
-(props) => {
-  const { idyllMarkup } = props;
-  return (
-    <IdyllDocument
-      ast={ compile(idyllMarkup) }
-      components={ components }
-      datasets={ {} } />
-  )
-}
-
-\`\`\`
-
-If not, you'll also need to install \`react\` and \`react-dom\`:
-
-\`\`\`sh
-$ npm i --save react react-dom
-\`\`\`
-
-and can embed it like this:
-
-\`\`\`javascript
-
-import React from 'react';
-import ReactDOM from 'react-dom';
-import IdyllDocument from 'idyll-document';
-import compile from 'idyll-compiler';
-import * as components from 'idyll-components';
-
-// You must provide idyllMarkup
-// and the container element (a DOM node).
-
-ReactDOM.render(
-  <IdyllDocument
-    ast={ compile(idyllMarkup) }
-    components={ components }
-    datasets={ {} } />,
-  containerElement
-)
-
-\`\`\`
-`
-
 
 export default ({ url }) => (
   <Layout url={ url }>
@@ -93,11 +27,10 @@ import compile from 'idyll-compiler';
 import * as components from 'idyll-components';
 
 // An example functional component
-(props) => {
-  const { idyllMarkup } = props;
+({ idyllMarkup }) => {
   return (
     <IdyllDocument
-      ast={ compile(idyllMarkup) }
+      markup={ idyllMarkup }
       components={ components }
       datasets={ {} } />;
   )
@@ -121,10 +54,9 @@ import * as components from 'idyll-components';
 
 // You must provide idyllMarkup
 // and the container element (a DOM node).
-
 ReactDOM.render(
   <IdyllDocument
-    ast={ compile(idyllMarkup) }
+    markup={ idyllMarkup }
     components={ components }
     datasets={ {} } />,
   containerElement
