+++ packages/idyll-document/package.json
@@ -31,6 +31,7 @@
     "change-case": "^3.0.1",
     "cross-env": "^5.0.5",
     "html-tags": "^2.0.0",
+    "idyll-compiler": "^2.5.7",
     "idyll-layouts": "^2.5.1",
     "idyll-themes": "^2.3.1",
     "object.entries": "^1.0.4",
