+++ packages/idyll-document/test/vars.js
@@ -5,19 +5,21 @@ import * as components from 'idyll-components';
 import IdyllDocument from '../src/';
 import ast from './fixtures/ast.json';
 
-let component;
+let component, idyllContext;
 
 const FAKE_DATA = 'FAKE DATA';
 
 beforeAll(() => {
-  component = mount(<IdyllDocument ast={ast} components={components} datasets={{myData: FAKE_DATA}} />);
+  component = mount(<IdyllDocument ast={ast} components={components} datasets={{myData: FAKE_DATA}} context={(ctx) => {
+    idyllContext = ctx;
+  }} />);
 })
 
 describe('Component state initialization', () => {
 
 
   it('creates the expected state', () => {
-    expect(component.state()).toEqual({
+    expect(idyllContext.data()).toEqual({
       x: 2,
       frequency: 1,
       xSquared: 4,
@@ -27,18 +29,18 @@ describe('Component state initialization', () => {
     });
   });
 
-  it('creates the expected derived vars', () => {
-    expect(component.instance().derivedVars).toEqual({
-      xSquared: {
-        value: 4,
-        update: expect.any(Function)
-      }
-    });
-  });
+  // it('creates the expected derived vars', () => {
+  //   expect(idyllContext.data().derivedVars).toEqual({
+  //     xSquared: {
+  //       value: 4,
+  //       update: expect.any(Function)
+  //     }
+  //   });
+  // });
 
-  it('can return the expected derived var values', () => {
-    expect(component.instance().getDerivedVars()).toEqual({xSquared: 4});
-  });
+  // it('can return the expected derived var values', () => {
+  //   expect(idyllContext.data()).toEqual({xSquared: 4});
+  // });
 
   it('renders expressions correctly before updates', () => {
     const displayComponents = component.findWhere((n) => {return n.type() === components.Display;});
@@ -88,9 +90,9 @@ describe('Component state initialization', () => {
 
 
   it('handles custom initial state', () => {
-    component = mount(<IdyllDocument ast={ast} initialState={{ x: 4 }} components={components} datasets={{myData: FAKE_DATA}} />);
+    component = mount(<IdyllDocument ast={ast} initialState={{ x: 4 }} components={components} datasets={{myData: FAKE_DATA}} context={(ctx) => idyllContext = ctx} />);
 
-    expect(component.state()).toEqual({
+    expect(idyllContext.data()).toEqual({
       x: 4,
       frequency: 1,
       xSquared: 16,
@@ -109,7 +111,7 @@ describe('Component state initialization', () => {
 
     updateProps({ value: 4 });
 
-    expect(component.instance().state).toEqual({
+    expect(idyllContext.data()).toEqual({
       x: 4,
       frequency: 1,
       xSquared: 16,
